Allan - Implemented JWT - Added JWT decorators, Protected routes, Configured JWT access token expiry period

+++ server/routes/documents_bp.py
@@ -1,8 +1,10 @@
 from flask import Blueprint, make_response, jsonify
 from flask_restful import Api, Resource, abort, reqparse
 from flask_marshmallow import Marshmallow
+from flask_jwt_extended import get_jwt_identity
 from serializer import documentSchema
 from models import db, Documents
+from auth_middleware import employee_required
 
 # create document blueprint
 document_bp = Blueprint('document_bp', __name__)
@@ -11,15 +13,18 @@ ma = Marshmallow(document_bp)
 api = Api(document_bp)
 
 
-parser = reqparse.RequestParser()
-parser.add_argument('employee_id', type=str, required=True,
-                    help="Document upload required")
-parser.add_argument('link_url', type=str, required=True,
-                    help="Document upload required")
-parser.add_argument('name', type=str, required=True,
-                    help="Name of document is required")
-parser.add_argument('type', type=str, required=True,
-                    help="Type of document is required")
+post_args = reqparse.RequestParser()
+post_args.add_argument('link_url', type=str, required=True,
+                       help="Document upload required")
+post_args.add_argument('name', type=str, required=True,
+                       help="Name of document is required")
+post_args.add_argument('type', type=str, required=True,
+                       help="Type of document is required")
+
+patch_args = reqparse.RequestParser()
+patch_args.add_argument('link_url', type=str)
+patch_args.add_argument('name', type=str)
+patch_args.add_argument('type', type=str)
 
 
 class AllDocuments(Resource):
@@ -29,11 +34,16 @@ class AllDocuments(Resource):
         response = make_response(jsonify(result), 200)
         return response
 
+    @employee_required()
     def post(self):
-        data = parser.parse_args()
+        current_user = get_jwt_identity()
+        data = post_args.parse_args()
 
         new_document = Documents(
-            **data
+            link_url=data["link_url"],
+            name=data['name'],
+            type=data['type'],
+            employee_id=current_user
         )
 
         db.session.add(new_document)
@@ -60,13 +70,18 @@ class DocumentById(Resource):
             response = make_response(jsonify(result), 200)
             return response
 
+    @employee_required()
     def patch(self, id):
+        current_user = get_jwt_identity()
         document = Documents.query.filter_by(id=id).first()
 
         if not document:
             abort(404, detail=f'document with id {id} does not exist')
 
-        data = parser.parse_args()
+        if document.employee_id != current_user:
+            abort(400, detail="Unauthorized request")
+
+        data = patch_args.parse_args()
         for key, value in data.items():
             if value is None:
                 continue
@@ -78,11 +93,16 @@ class DocumentById(Resource):
 
         return response
 
+    @employee_required()
     def delete(self):
+        current_user = get_jwt_identity()
         document = Documents.query.filter_by(id=id).first()
 
         if not document:
             abort(404, detail=f'document with id {id} does not exist')
+
+        if document.employee_id != current_user:
+            abort(400, detail="Unauthorized request")
 
         db.session.delete(document)
         db.session.commit()
